Remove the dependencies of lz4 stuffs from bitshuffle cores
