+++ src/bshuf_h5filter.c
@@ -11,7 +11,7 @@
  */
 
 #include "bshuf_h5filter.h"
-#include "bitshuffle.h"
+#include "bshuf_lz4.h"
 
 
 #define PUSH_ERR(func, minor, str)                                      \
